perf(core): optimize query hot path, telemetry, and solver ingress
- Add cache fast-paths to MassUtility.Capacity and MassCapacityStatWorker,
  bypassing baseline resolution, locks, and Pawn.BodySize on cache hits.
- Replace main-thread atomic telemetry counters with gated plain increments
  to eliminate bus-lock stalls and measurement probe overhead.
- Cache HediffStage physiological impacts in MedicalClassifier for O(1) lookups
  during live pathology sweeps.
- Precompute dynamic cache expiry sizing to eliminate repeated Mathf.Sqrt calls.
- Deduplicate pawn consciousness evaluations across ingress and solver passes.

+++ Source/Settings/SettingsDefaults.cs
@@ -192,7 +192,7 @@ namespace OverHaulers
         public const int DefaultCacheMinFastPathTicks = 5;
 
         // Dynamic square-root scaling factor (K) for population-based cache expiry.
-        public const float DynamicCacheExpiryScaleK = 5.0f;
+        public const int DynamicCacheExpiryScaleK = 5;
 
         // Minimum tick interval between reactive cache invalidations for a single pawn during combat damage bursts (15 ticks = ~0.25s).
         public const int InvalidationDebounceTicks = 15;
